test: move concurrency race tests into their feature test suites
Relocate the register, accept-invite and create-channel race tests from
the dedicated ConcurrentCreationRaceTests file into AuthEndpointsTests,
AcceptInviteEndpointTests and CreateChannelTests, matching how the
OpenConversation race test lives in ConversationEndpointsTests.
